Changed font sizes to be ratios of the current viewport height.

+++ include/arch/asset_management/font.h
@@ -3,6 +3,6 @@
 
 #include "asset_management/asset_manager_types.h"
 
-int32 loadFont(const char *name, uint32 size);
-Font* getFont(const char *name, uint32 size);
+int32 loadFont(const char *name, real32 size);
+Font* getFont(const char *name, real32 size);
 void freeFont(Font *font);
