+++ resources/scripts/components/font.lua
@@ -2,7 +2,7 @@ ffi.cdef[[
 typedef struct font_component_t
 {
 	char name[64];
-	uint32 size;
+	real32 size;
 	kmVec4 color;
 } FontComponent;
 ]]
